Add files via upload
This is my second training project on working with Spring. The above project was compiled on Maven using Spring Boot 3. The point of the project is the work of the nursery in recording seedlings (quantity, cost, variety, price, year of planting). OneToMany dependency was used. The variety of co-parents is selected from Enum. The time when records were created and modified is also tracked. 
The project is fully working and tested. 
The project is covered with unit and integration tests.